BroadcastReceiver and CharacteristicSorter
created two new classes, one for the sorting next for the broadcastreceiver. Sortingclass got explained by last commit. Now the BroadcastReiver object is separate from the Activity, nevertheless the behavior if intent.getAction fulfills a condition calls methods for the visuall appearance aswell. PROBABLY need to refactore anothertime.  Nevertheless this Broadcaster can be customized for all activities, therefore it needs to be extended with proper behavior and a Filterobject. Definition of filterobjects can be lookedup as an example from already existing one.  #20 seems to fit. NEED TO KNOW: to make those visualupdates possible those methods are not set to package-private NOT private anymore...AND needed to implement some getter methods to operate on activityspecific objects like bluetoothLeService